+++ tests/Collection/ConfigurationCollectionTest.php
@@ -18,10 +18,9 @@ class ConfigurationCollectionTest extends ContaoTestCase
 {
     public function createTestInstance(array $parameters = [])
     {
-        $bundleConfig = $parameters['bundleConfig'] ?? [];
         $extensionCollection = $parameters['extensionCollection'] ?? $this->createMock(ExtensionCollection::class);
 
-        return new ConfigurationCollection($bundleConfig, $extensionCollection);
+        return new ConfigurationCollection($extensionCollection);
     }
 
     public function testGetJsEntries()
@@ -47,23 +46,13 @@ class ConfigurationCollectionTest extends ContaoTestCase
         $this->assertCount(1, $instance->getJsEntries(['array' => true]));
         $this->assertIsArray($instance->getJsEntries(['array' => true])[0]);
 
-        $bundleConfig = [
-            'js_entries' => [
-                ['name' => 'bundle2', 'file' => 'bundle2', 'requires_css' => true],
-                ['name' => 'bundle3', 'file' => 'bundle3', 'head' => true],
-            ],
-        ];
-
         $instance = $this->createTestInstance([
             'extensionCollection' => $extensionCollection,
-            'bundleConfig' => $bundleConfig,
         ]);
-        $this->assertCount(3, $instance->getJsEntries());
+        $this->assertCount(1, $instance->getJsEntries());
         $this->assertInstanceOf(EncoreEntry::class, $instance->getJsEntries()[0]);
-        $this->assertInstanceOf(EncoreEntry::class, $instance->getJsEntries()[1]);
 
-        $this->assertCount(3, $instance->getJsEntries(['array' => true]));
+        $this->assertCount(1, $instance->getJsEntries(['array' => true]));
         $this->assertIsArray($instance->getJsEntries(['array' => true])[0]);
-        $this->assertIsArray($instance->getJsEntries(['array' => true])[1]);
     }
 }